refactor: auto-split test_integration_roundtrip.py
